requirements.txt
google-auth==2.41.1
gspread==6.2.1
pandas==2.3.3
PyGithub==2.8.1
tqdm==4.67.1
PyYAML==6.0.3
pytest==9.0.1
huggingface_hub==1.2.3